refactor(process): sample-processor replaced
