+++ packages/botonic-nlp/src/tasks/ner/process/types.ts
@@ -1,6 +1,7 @@
 import { Tensor2D, Tensor3D } from '@tensorflow/tfjs-node'
 
-export type Set = { x: Tensor2D; y: Tensor3D }
+export type InputData = Tensor2D
+export type OutputData = Tensor3D
 
 export type Prediction = { label: string; confidence: number }
 export type Entity = {
