--- packages/botonic-nlp/tests/tasks/ner/process/ner-sample-processor.test.ts
@@ -1,19 +1,0 @@
-import { NerSampleProcessor } from '../../../../src/tasks/ner/process/sample-processor'
-import * as helper from '../../../helpers/tasks/ner/helper'
-
-describe('NER Sample Processor', () => {
-  test('Process sample', () => {
-    const processor = new NerSampleProcessor(
-      helper.preprocessor,
-      helper.sequenceCodifier,
-      helper.entitiesCodifier
-    )
-    expect(
-      processor
-        .process([
-          { class: '', entities: [], text: 'I love this leather jacket' },
-        ])
-        .x.arraySync()
-    ).toEqual([[2, 28, 20, 17, 0, 0, 0, 0, 0, 0, 0, 0]])
-  })
-})
